Add ManageIQ::FileStorage::TypeRegistry
Starts the process of making this library more "pluggable" by
introducing a TypeRegistry that can be used to add new FileStorage
classes.

Functionally, it ends up with the same public interface as before, but
is able to be extended without requiring this project to being the only
source of plugins.  Additionally, it individual or a subset of file
storage plugins can now be loaded instead of all having to be loaded for
to function, since each are registered on class definition.