Add "docker-machine" support through Vagrant
While the `docker-machine` utility exists for ease of spinning up a
local docker host, it has a deficiency in it's implementation which
makes it relatively difficult to modify the VM it is running docker on.

This is specifically an issue for the use case of this project where to
run `nfsd` in a docker container, the `nfs` and `nfsd` kernel modules
are required, which the `boot2docker` VM does not have included.

The implementation here attempts to imitate the functionality of
`docker-machine` + `docker-compose` as much as possible, while
streamlining it's setup via `rake`.

** Rakelib **

Intentionally left empty at the moment, since the default behaviour of
`rake` will load any `*.rake` files in the `rakelib/` dir.

** rakelib/docker.rake **

Configures tasks to work with "docker-machine"

- 'docker:machine:up':    Spins up a dockerd VM, listening on https
- 'docker:machine:env':   Sets environment vars for use with dockerd VM
- 'docker:machine:down':  Destroys the above VM
- 'docker:source_env':    Prints 'DOCKER_*' env vars and shell usage
- 'docker:compose:*':     `docker-compose` tasks, configured to above VM

** Vagrantfile **

The vagrant VM that simulates a `docker-machine` VM.

- Generates/Configures TLS certs
- Installs, configures, and runs dockerd

** mkdockercerts.rb **

Attempts replicate the TLS cert generation process found in
`docker-machine`.  Generates `ca`, `client`, and `server` certs, and
puts them in the `.docker-certs` dir relative to the `Vagrantfile`

** docker-compose.yml **

Currently configures a `NFS` server container for testing.  More types
are expected to be added in the future.