Add TestfileServer
This is a helper script intended to provide a standardized way of
creating test files through a simple HTTP API.

Since NFS can't mount encrypted volumes (which is what you would get
mostly when sharing a folder via VirtualBox), this is the next best
solution.  The process is pretty light weight, so it shouldn't provide
much of a strain of the VM, and it really is only intended for local dev
and CI (CI only to maintain the same interface in the specs).

There are two HTTP actions per configured mount type:

* POST:    Create a new test file. If "size" is passed, create it with
           that size
* DELETE:  Remove the contents of the configured directory

The name of the new temp file that is created is returned on a POST as
plain text, so if that is necessary for the spec, that is provided, and
the default size is 0 when it is created.

On the Vagrant VM, the service is controlled via openrc and is scheduled
as part of the bootstrap provisioner.

Of note:  While "tmpdir" is included here, the code only makes use of
the ruby `Tmpname` module in this code.  The reason for this is if the
ruby server is terminated and `Tempfile` instances were being used, they
would be deleted on exit of the ruby process.  This seems to be less
than ideal for debugging purposes, so regular files are used, but
`Tempname` is still used to generate uniq names per call.