[spec/support/custom_matchers] Make generic part 1
Changes the ftp specific custom matchers for rspec to have names that
are more generic and can eventually be used for other file storage
providers.

This current change just handles the rename before making more sweeping
changes to the interface.  Currently only works with FTP, but doing it
this way makes the diff nicer.