[WIP] Using ruby based DockerContainerManager
Instead of something like `docker-compose`

Requires docker-api gem.

Pros:
- just a ruby gem, no C extensions
- easy interface to upload files (no extra service/ruby needed)
- removes need for docker-compose/docker dependencies
- removes need for http server on VM
- removes need for extra http client for test drivers
- note a terrible amount of extra code need for manager

Cons:
- Extra code is needed for manager
- DSL is custom, and is one more thing to learn
- Probably still requires a bit of code re-working for CI

Regarding the last point with CI:

The intent with the `docker-compose` route is that there would be a few
config variables that would be needed to change how the specs are run
when on a linux host versus macOS/win (when using "Vagrant
docker-machine"), and most likely `docker-compose` would be triggered to
run just before the test suite on CI.  This means any `docker pull` that
would be necessary would happen prior to `rake` being called.

For this approach, having a way to pre-pull the containers on the CI
host prior to doing running `rake` would probably be idea, just so specs
are slow when a new image is spun up.

The intent with this is that the `with_XXX_server` contexts would only
spin up the containers for the contexts they are needed, and spin them
down afterwards (similar to what is being done for FTP).  To avoid
having to weight for a `docker pull` in the middle of the specs, some
way of getting an inventory of the base container images that will need
to be made available for rake.  The current interface doesn't allow for
that, but it wouldn't be too hard (a .yaml file might be just fine) and
how it is done just needs to be solidified.