Add TestfileServerAPIClient
This is a api client driver for the TestfileServer created in the
previous commit.  It allows for a WithFileStorageInterface::Driver
interface for working with the filesystem directly without needing to
connect via `mount` to do so to create existing files and query what is
currently there.

This will be used when testing the mounted file systems via docker.