[WIP][PLACE HOLDER] Intial commit
This is currently a place holder commit, and this commit and branch will
be rebased in the future.

A working branch (most likely named `historical_master`) will be pushed
shortly, and progress on the extraction and refactoring of this base can
be followed there in the mean time, though expect that branch to be
volatile until that process is complete.