Add sudo_helper.rb
This is a helper for rspec to prime sudo for use with rspec, so auth for
rspec can happen prior to the specs being run.  This gives a much more
obvious behavior to the user instead of the auth being requested partway
through the specs.

In addition, spec files that require sudo usage can check for it via the
`RSpec.configuration.with_sudo` setting that will be set after priming
sudo.